+++ app/views/observations/downloads/new.html.erb
@@ -1,4 +1,5 @@
 <%
+# The form partial is also called by species_lists/downloads
   @title = :download_observations_title.t
 
   # tabs = [
@@ -7,58 +8,6 @@
   # @tabsets = { right: draw_tab_set(tabs) }
 %>
 
-<%= form_with(url: observations_downloads_path(q: get_query_param)) do |f| %>
-  <p><%= :download_observations_format.t %>:</p>
-  <ul type="none">
-    <li>
-      <%= radio_button_tag(:format, :raw, @format == "raw") %>
-      <%= :download_observations_raw.t %>
-    </li>
-    <li>
-      <%= radio_button_tag(:format, :adolf, @format == "adolf") %>
-      <%= :download_observations_adolf.t %>
-    </li>
-    <li>
-      <%= radio_button_tag(:format, :darwin, @format == "darwin") %>
-      <%= :download_observations_darwin.t %>
-      (<%= link_to(:download_observations_what_is_this.t,
-                   "http://rs.tdwg.org/dwc/index.htm",
-                   target: "_new") %>)
-    </li>
-    <li>
-      <%= radio_button_tag(:format, :symbiota, @format == "symbiota") %>
-      <%= :download_observations_symbiota.t %>
-    </li>
-    <li>
-      <%= radio_button_tag(:format, :fundis, @format == "fundis") %>
-      <%= :download_observations_fundis.t %>
-    </li>
-  </ul>
+<%= render(partial: "observations/downloads/form",
+           locals: { query_param: get_query_param }) %>
 
-  <p><%= :download_observations_encoding.t %>:</p>
-  <ul type="none">
-    <li>
-      <%= radio_button_tag(:encoding, "ASCII", @encoding == "ASCII") %>
-      <%= :download_observations_ascii.t %>
-    </li>
-    <li>
-      <%= radio_button_tag(:encoding, "WINDOWS-1252", @encoding == "WINDOWS-1252") %>
-      <%= :download_observations_windows.t %>
-    </li>
-    <li>
-      <%= radio_button_tag(:encoding, "UTF-8", @encoding == "UTF-8") %>
-      <%= :download_observations_utf8.t %>
-    </li>
-    <li>
-      <%= radio_button_tag(:encoding, "UTF-16", @encoding == "UTF-16") %>
-      <%= :download_observations_utf16.t %>
-    </li>
-  </ul>
-
-  <%= submit_tag(:DOWNLOAD.l, class: "btn btn-default") %>
-  <%= submit_tag(:CANCEL.l, class: "btn btn-default") %>
-
-  <p class="mt-5"><%= :download_observations_print_labels_header.t %>:</p>
-  <%= submit_tag(:download_observations_print_labels.l, class: "btn btn-default") %>
-
-<% end %>
