+++ app/views/observations/show/_lists.html.erb
@@ -7,8 +7,8 @@
       <% observation.species_lists.each do |list| %>
         <tr>
           <td style="border-top:0">
-            <%= link_to(list.format_name.t, controller: "/species_list",
-                        action: :show_species_list, id: list.id) %>
+            <%= link_to(list.format_name.t, controller: "/species_lists",
+                        action: :show, id: list.id) %>
             <%= if check_permission(list)
               url = {
                 controller: "/species_list",
