+++ app/views/species_lists/new.html.erb
@@ -2,7 +2,7 @@
   @title = :species_list_create_title.t
 
   tabs = [
-    link_to(:name_lister_title.t, action: :name_lister)
+    link_to(:name_lister_title.t, new_species_list_name_lister)
   ]
   @tabsets = { right: draw_tab_set(tabs) }
   @container = :wide
