+++ app/views/species_lists/projects/edit.html.erb
@@ -5,13 +5,18 @@
     link_to(:cancel_and_show.t(type: :species_list), @list.show_link_args)
   ]
   @tabsets = { right: draw_tab_set(tabs) }
+
+  action = { controller: "/species_lists/projects",
+             action: :update, id: @list.id }
 %>
 
 <div class="help-block mt-3">
   <%= :species_list_projects_help.tp %>
 </div>
 
-<%= form_tag(id: @list.id) %>
+<%= form_with(url: action, method: :put,
+              id: "species_list_projects_form") do |f| %>
+
   <div class="form-group form-inline mt-3">
     <b><%= :species_list_projects_which_objects.t %></b><br/>
     <%= check_box_tag(:objects_list, 1, @object_states[:list],
@@ -35,8 +40,9 @@
   </div>
 
   <div class="text-center mt-3">
-    <%= submit_tag(:ATTACH.l, class: "btn btn-default") %>
+    <%= f.submit(:ATTACH.l, class: "btn btn-default") %>
     &nbsp;
-    <%= submit_tag(:REMOVE.l, class: "btn btn-default") %>
+    <%= f.submit(:REMOVE.l, class: "btn btn-default") %>
   </div><!-- .text-center -->
-</form>
+
+<% end %>
